Refactor ProjectMeta: expose flags directly instead
We tried to merge the flags into the repository settings (where they should
belong), but this leads to an extremely ugly implementation when trying to do
the same for packages, as they can no longer be separated from the
project. Also, it leads to nasty issues if you first try to setup your flags and
then configure the repositories, as then the flags will be silently erased.