Fix data race in makeApiCall
xml2js' Parser.parseStringPromise is not "thread safe": it appears to store
internal state in the Parser object and is thus not safe to be called in a
concurrently.

The documentation sorta-kinda mentions this here:
https://github.com/Leonidas-from-XIV/node-xml2js#parsing-multiple-files

To be safe, we now create a new Parser & Builder for each request, so that no
state can be shared across requests.