Use content-type application/octet-stream & drop xml header
OBS expects the received body of PUT and POST requests to have this content
type and since it chokes on chunked transfers, we must turn these off explicitly
by specifying a content length.
During experimenting I also figured out that one remove the xml headline
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
Since it is not required and not used by neither osc nor obs, we drop it for
better compatibility.