Use a temporary directory instead of mock-fs in package tests
For some reason the mock-fs module breaks the xmlbuilder in this case by
overriding the fs module... Node is then not able to load some files on which
xml2js depends and the test fails for weird reasons. Using a temporary directory
is not as "clean", but it at least works reliably.