Fix uncaught exceptions in runProcess when spawned command does not exist
When the spawned command does not exist, then we will not get a non-zero return
value from spawn, but instead the `error` event is emitted, which we did not
handle. When that event is unhandled and an error occurs, then runProcess throws
uncatchable exceptions which lead to really weird failures downstream.